% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastring.R
\docType{data}
\name{datastring}
\alias{datastring}
\title{A string vector that simulates a Fasta input.}
\format{A string vector with six strings.}
\usage{
datastring
}
\description{
A dataset containing a big sequences of nucleic acids for quick
testing the package.
}
\keyword{datasets}
