% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthbase.R
\name{orthbase}
\alias{orthbase}
\title{Generate a orthonormal matrix (\code{lin} , \code{col})}
\usage{
orthbase(lin, col)
}
\arguments{
\item{lin}{Number of rows in the desired matrix}

\item{col}{Number of columns in the desired matrix}
}
\value{
A orthonormal matrix in a specified size, \code{lin} by \code{col}.
}
\description{
Generate a orthonormal matrix in a specified
size, \code{lin} by \code{col}.
}
\examples{
orthbase(160000, 10)

lin <- 160000
col <- 10
orthbase(lin = lin, col = col)


}
\seealso{
\code{\link[SVect]{SWeeP}}, \code{\link[pracma]{orth}}
}
\author{
Roberto Raitz, et. al.
}
