% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sWeeP.R
\name{sWeeP}
\alias{sWeeP}
\alias{sWeeP,character-method}
\alias{sWeeP,AAStringSet-method}
\title{A vectorial comparative method to amino acid sequence.}
\usage{
sWeeP(xfas, baseMatrix)

\S4method{sWeeP}{character}(xfas, baseMatrix)

\S4method{sWeeP}{AAStringSet}(xfas, baseMatrix)
}
\arguments{
\item{xfas}{A AAStringSet or a FASTA format file}

\item{baseMatrix}{A orthonormal matrix with 160.000 coordinates}
}
\value{
A matrix resulted by the projection of the sequences in \code{xfas}
in the \code{baseMatrix} matrix
}
\description{
The "Spaced Words Projection (SWeeP)" is a method for
representing biological sequences using compact vectors. SWeeP uses the
spacedwords concept by scanning sequences and generating indices
to create a higherdimensional matrix of occurrences that is later projected into a
smaller randomly oriented orthonormal base (PIERRI, 2019). This way the resulting
matrix  will conserve the comparational data but will have a selectable size
}
\details{
The SWeeP method was developed to favor the comparison between
complete proteomic sequences and to assist in machine learning analyzes. This
method is based on the concept of spaced words, which are used to scan
biological sequences and project them into  matrix of occurrences, favoring the manipulation
of the data. The sWeeP function can project a matrix n by m, where n is the number of
sequences in the analized \code{xfas} and m is the number of columns
in \code{baseMatrix} matrix.
}
\examples{
baseMatrix <- orthBase(160000,10)
path <- system.file(package = "rSWeeP", "extdata", "exdna.fas")
return <- sWeeP(path,baseMatrix)
distancia <- dist(return, method = "euclidean")
tree <- hclust(distancia, method="ward.D")
plot(tree, hang = -1, cex = 1)
}
\references{
Pierri,C. R. et al. SWeeP: Representing large biological
sequences data sets in compact vectors. Scientific Reports, accepted in
December 2019.doi: 10.1038/s41598-019-55627-4.
}
\seealso{
\code{\link[rSWeeP:orthBase]{orthBase}}
}
\author{
Danrley R. Fernandes.
}
