% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWeeP.R
\name{SWeeP}
\alias{SWeeP}
\title{A vectorial comparative method to amino acid sequence.}
\usage{
SWeeP(file, orthBase)
}
\arguments{
\item{file}{A FASTA format file}

\item{orthBase}{A orthonormal matrix with 160.000 rows}
}
\value{
A matrix resulted by the projection of the sequences in \code{file}
in the \code{orthBase} matrix
}
\description{
The "Spaced Words Projection (SWeeP)" is a method for
representing biological sequences using relativety, it uses the
spacedwords concept by scanning sequences and generating indices
to create a higherdimensional vector that is later projected into a
smaller randomly oriented orthonormal base (PIERRI, 2019)
}
\details{
The SWeeP method was developed to favor the comparison between
complete genomic sequences and to assist in phylogenetic analyzes. This
method is based on the concept of sparse words, which are used to scan
biological sequences and turn them into vectors, favoring the manipulation
of the data. It can project a matrix n by m, where n is the number of
sequences in the analized \code{file} and m is the number of columns
in \code{orthBase} matrix.
}
\examples{
baseMatrix <- orthbase(160000,10)
data(datastring)
cat(datastring,file = "exdna.fas", sep = "\\n")
return <- SWeeP("exdna.fas",baseMatrix)
distancia <- dist(return, method = "euclidean")
tree <- hclust(distancia, method="ward.D")
plot(tree, hang = -1, cex = 1)
}
\references{
Pierri,C. R. et al. SWeeP: Representing large biological
sequences data sets in compact vectors. (2019).
}
\seealso{
\code{\link[rSWeeP]{orthbase}}
}
\author{
Roberto Raitz, et. al.
}
